+++ app/types/domain/MeetupGroup.ts
@@ -1,0 +1,6 @@
+import { MeetupEvent } from './MeetupEvent';
+
+export type MeetupGroup = {
+  id: string;
+  pastEvents: MeetupEvent[] | null;
+};